Cleaned up some files and folders, changed stuff to smart pointers in areas

+++ GameEngine/Source/Engine/Vector3.h
@@ -44,7 +44,7 @@ namespace sr
 		Vector3 operator /= (float s) { this->x /= s; this->y /= s; this->z /= s; return *this; }
 
 		float LengthSqr() const { return (float)(pow(x, 2) + pow(y, 2) + pow(z, 2)); }
-		float Length() const { return sqrt(LengthSqr()); }
+		float Length() const { return (float)sqrt(LengthSqr()); }
 		Vector3 Normalize() const { return (*this) / Length(); }
 		float Dot(const Vector3& v) const { return (this->x * v.x) + (this->y * v.y) + (this->z * v.z); }
 	};
